+++ dart/geobase/lib/src/common/reference/coord_ref_sys.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
@@ -96,12 +96,23 @@ class CoordRefSys {
   bool swapXY({GeoRepresentation? logic}) =>
       CoordRefSysResolver.registry.swapXY(id, logic: logic);
 
+  /// Returns an EPSG code according to the common `EPSG:{code}` template
+  /// for [id] if the coordinate reference system is recognized by the
+  /// [EPSG register](https://epsg.org/).
+  ///
+  /// For example for `http://www.opengis.net/def/crs/EPSG/0/4326` (WGS 84
+  /// latitude/longitude) this getter returns `4326` as an integer, but for
+  /// `http://www.opengis.net/def/crs/OGC/1.3/CRS84` this returns null as CRS84
+  /// (WGS 84 longitude/latitude) do not have an exact corresponding identifier
+  /// in the EPSG register.
+  int? get epsgCode => CoordRefSysResolver.registry.epsgCode(id);
+
   /// Returns an EPSG identifier according to the common `EPSG:{code}` template
   /// for [id] if the coordinate reference system is recognized by the
   /// [EPSG register](https://epsg.org/).
   ///
   /// For example for `http://www.opengis.net/def/crs/EPSG/0/4326` (WGS 84
-  /// latitude/longitude) this getter returns `EPSG:4326`, but for
+  /// latitude/longitude) this getter returns `EPSG:4326` as a string, but for
   /// `http://www.opengis.net/def/crs/OGC/1.3/CRS84` this returns null as CRS84
   /// (WGS 84 longitude/latitude) do not have an exact corresponding identifier
   /// in the EPSG register.