+++ dart/geobase/CHANGELOG.md
@@ -10,6 +10,7 @@ NOTE: Version 1.1.0 currently under development (1.1.0-dev.0).
 * [Add EWKB encoding and decoding support on geobase #165](https://github.com/navibyte/geospatial/issues/165)
   * Default endianess to Endian.little in WKB and some docs updated. 
   * Support EWKB dimensionality flags for z and m.
+  * CoordRefSys to support getting EPSG code as an integer.
 
 ## 1.0.2